replaced fetch api sequences with axios, relocated starknet logic
