+++ frontend/src/pages/forms/Form.jsx
@@ -2,8 +2,8 @@ import './form.css';
 import React, { useEffect, useState } from 'react';
 import TokenSelector from '../../components/TokenSelector';
 import MultiplierSelector from '../../components/MultiplierSelector';
-import { connectWallet, getBalances } from '../../utils/wallet';
-import { handleTransaction } from '../../utils/transaction';
+import { connectWallet, getBalances } from '../../services/wallet';
+import { handleTransaction } from '../../services/transaction';
 import Spinner from '../../components/spinner/Spinner';
 import { ReactComponent as ETH } from '../../assets/icons/ethereum.svg';
 import { ReactComponent as USDC } from '../../assets/icons/borrow_usdc.svg';
